clean up after synchronous review
