+++ great_expectations/data_context/data_context/abstract_data_context.py
@@ -1,22 +1,31 @@
 import configparser
 import copy
+import errno
 import json
 import logging
 import os
-from abc import abstractmethod
-from typing import Dict, Mapping, Optional, Union
+from abc import ABC, abstractmethod
+from typing import Dict, Mapping, Optional, Union, cast
 
-from great_expectations.core.config_peer import ConfigPeer
+from ruamel.yaml import YAML
+
 from great_expectations.data_context.types.base import (
     DataContextConfig,
     anonymizedUsageStatisticsSchema,
 )
-from great_expectations.data_context.util import substitute_all_config_variables
+from great_expectations.data_context.util import (
+    substitute_all_config_variables,
+    substitute_config_variable,
+)
 
 logger = logging.getLogger(__name__)
 
+yaml = YAML()
+yaml.indent(mapping=2, sequence=4, offset=2)
+yaml.default_flow_style = False
 
-class AbstractDataContext(ConfigPeer):
+
+class AbstractDataContext(ABC):
     """
     Base class for all DataContexts that contain all context-agnostic data context operations.
 
@@ -121,9 +130,36 @@ class AbstractDataContext(ConfigPeer):
 
     def _load_config_variables(self) -> Dict:
         """
-        Get all config variables from default location (empty in the case of AbstractDataContext) Overriden by FileDataContext
+        Get all config variables from the default location. For Data Contexts in GE Cloud mode, config variables
+        have already been interpolated before being sent from the Cloud API.
+
         """
-        return {}
+        config_variables_file_path: str = cast(
+            DataContextConfig, self._project_config
+        ).config_variables_file_path
+        if config_variables_file_path:
+            try:
+                # If the user specifies the config variable path with an environment variable, we want to substitute it
+                defined_path: str = substitute_config_variable(
+                    config_variables_file_path, dict(os.environ)
+                )
+                if not os.path.isabs(defined_path) and hasattr(self, "root_directory"):
+                    # A BaseDataContext will not have a root directory; in that case use the current directory
+                    # for any non-absolute path
+                    root_directory: str = self.root_directory or os.curdir
+                else:
+                    root_directory: str = ""
+                var_path = os.path.join(root_directory, defined_path)
+                with open(var_path) as config_variables_file:
+                    res = dict(yaml.load(config_variables_file))
+                    return res or {}
+            except OSError as e:
+                if e.errno != errno.ENOENT:
+                    raise
+                logger.debug("Generating empty config variables file.")
+                return {}
+        else:
+            return {}
 
     @staticmethod
     def _get_global_config_value(
@@ -260,13 +296,3 @@ class AbstractDataContext(ConfigPeer):
         }
 
         return substitutions
-
-    @property
-    def config(self) -> DataContextConfig:
-        """
-        Required because of ConfigPeer
-
-        Returns:
-            DataContextConfig
-        """
-        return self._project_config
