+++ great_expectations/data_context/data_context/file_data_context.py
@@ -1,7 +1,5 @@
-import errno
 import logging
-import os
-from typing import Mapping, Optional, Union, cast
+from typing import Mapping, Optional, Union
 
 from ruamel.yaml import YAML
 
@@ -12,7 +10,6 @@ from great_expectations.data_context.types.base import DataContextConfig
 from great_expectations.data_context.types.data_context_variables import (
     FileDataContextVariables,
 )
-from great_expectations.data_context.util import substitute_config_variable
 
 logger = logging.getLogger(__name__)
 
@@ -53,38 +50,6 @@ class FileDataContext(AbstractDataContext):
 
     def _init_variables(self) -> FileDataContextVariables:
         raise NotImplementedError
-
-    def _load_config_variables(self) -> dict:
-        """
-        Get all config variables from the default location. For Data Contexts in GE Cloud mode, config variables
-        have already been interpolated before being sent from the Cloud API.
-        """
-        config_variables_file_path: str = cast(
-            DataContextConfig, self.get_config()
-        ).config_variables_file_path
-        if config_variables_file_path:
-            try:
-                # If the user specifies the config variable path with an environment variable, we want to substitute it
-                defined_path: str = substitute_config_variable(
-                    config_variables_file_path, dict(os.environ)
-                )
-                if not os.path.isabs(defined_path):
-                    # A BaseDataContext will not have a root directory; in that case use the current directory
-                    # for any non-absolute path
-                    root_directory: str = self.root_directory or os.curdir
-                else:
-                    root_directory: str = ""
-                var_path = os.path.join(root_directory, defined_path)
-                with open(var_path) as config_variables_file:
-                    res = dict(yaml.load(config_variables_file))
-                    return res or {}
-            except OSError as e:
-                if e.errno != errno.ENOENT:
-                    raise
-                logger.debug("Generating empty config variables file.")
-                return {}
-        else:
-            return {}
 
     @property
     def root_directory(self) -> str:
